client-samples/python/websockets/requirements.txt
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
idna==3.4
msal==1.25.0
pycparser==2.21
PyJWT==2.8.0
requests==2.31.0
urllib3==2.1.0
websocket-client==1.6.4
black==23.11.0